refactored search bar
