+++ package.json
@@ -13,6 +13,10 @@
     "@reduxjs/toolkit": "^1.6.2",
     "babel-plugin-transform-inline-environment-variables": "^0.4.3",
     "expo": "~42.0.1",
+    "expo-camera": "~11.2.2",
+    "expo-image-picker": "~10.2.3",
+    "expo-media-library": "~12.1.2",
+    "expo-permissions": "~12.1.1",
     "expo-splash-screen": "~0.11.2",
     "expo-status-bar": "~1.0.4",
     "expo-updates": "~0.8.1",
@@ -21,20 +25,18 @@
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "~0.63.4",
+    "react-native-event-source": "^1.1.0",
     "react-native-gesture-handler": "~1.10.2",
     "react-native-reanimated": "~2.2.0",
     "react-native-screens": "~3.4.0",
+    "react-native-sse": "^1.1.0",
     "react-native-star-rating": "^1.1.0",
     "react-native-unimodules": "0.14.10",
     "react-native-video": "^5.2.0",
     "react-native-web": "~0.13.12",
     "react-redux": "^7.2.6",
     "redux": "^4.0.5",
-    "yup": "^0.32.11",
-    "expo-image-picker": "~10.2.3",
-    "expo-permissions": "~12.1.1",
-    "expo-camera": "~11.2.2",
-    "expo-media-library": "~12.1.2"
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
